+++ public/script/oldGardens.js
@@ -42,16 +42,16 @@ editBtn.forEach(btn=>{
 
 deleteBtn.forEach(btn=>{
     btn.addEventListener('click', e => {
-        console.log(btn.value)
-        console.log(`/api/garden/${btn.value}`)
-        fetch(`/api/garden/${btn.value}`,{
-            method:"DELETE"
-        }).then(res=>{
-            if(res.ok){
-               location.reload()
-            } else {
-                alert("trumpet sound")
-            }
-        })
-    })
+        if (window.confirm("This action cannot be undone")) {
+            fetch(`/api/garden/${btn.value}`,{
+                method:"DELETE"
+            }).then(res=>{
+                if(res.ok){
+                   location.reload()
+                } else {
+                    alert("trumpet sound")
+                }
+            });
+        }
+    });
 })
